+++ src/ObjectsManager/Object.hpp
@@ -26,16 +26,15 @@ enum class ObjectState { Unchanged, Modified, Added, Removed };
 
 class Object {
  public:
-  Object(const fs::path &filePath, const std::string &objName, ObjectType type, uintmax_t startLine,
-         uintmax_t startColumn, uintmax_t endLine, uintmax_t endColumn, const std::string &rawComment,
-         const std::string &debrief, const std::vector<std::string> &arguments, const std::string &returnType,
-         ObjectState state);
+  Object(const fs::path &filePath, const std::string &objName, ObjectType type, size_t startLine, size_t startColumn,
+         size_t endLine, size_t endColumn, const std::string &rawComment, const std::string &debrief,
+         const std::vector<std::string> &arguments, const std::string &returnType, ObjectState state);
   Object(const json::json &j);
   ~Object() = default;
   auto operator==(const Object &other) const -> bool;
   auto isValid() const -> bool;
   auto setState(ObjectState state) -> void;
-  auto setOverloadIndex(uintmax_t index) -> void;
+  auto setOverloadIndex(size_t index) -> void;
   auto updateObject(const Object &other) -> void;
   auto getState() const -> ObjectState;
   auto getStateAsString() const -> std::string;
@@ -51,11 +50,11 @@ class Object {
   fs::path filePath_;
   std::string name_;
   ObjectType type_;
-  uintmax_t overloadIndex_;
-  uintmax_t startLine_;
-  uintmax_t startColumn_;
-  uintmax_t endLine_;
-  uintmax_t endColumn_;
+  size_t overloadIndex_;
+  size_t startLine_;
+  size_t startColumn_;
+  size_t endLine_;
+  size_t endColumn_;
   std::string rawComment_;
   std::string debrief_;
   std::vector<std::string> arguments_;
